Implement Step 4
